+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changes
 0.3.4 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Use regex to parse file status reponse instead of complicated string split/join
 
 
 0.3.3 (2013-01-28)